Final Commit
Ready for Review. Renamed the index.html. Filled in the comments on all of the documents. Loaded in the scripts that need to be loaded in. Removed the _txt from all of the scripts. Added in a script to store all of the array data. Added a script which makes the html to display all of the items defined in that script element. The orderTotal is also calculated within the script. Added in comments to all of the code. Added in my name and the date in all of the top comments also.